chore: show specific local storage error message
