+++ src/components/ErrorBoundary/index.tsx
@@ -1,42 +1,7 @@
-import { Trans } from '@lingui/macro'
 import { captureException } from '@sentry/react'
 import React, { ErrorInfo, PropsWithChildren } from 'react'
-import { Text } from 'rebass'
-import styled from 'styled-components/macro'
-import { UAParser } from 'ua-parser-js'
 
-import { ButtonPrimary } from 'components/Button'
-import { AutoColumn } from 'components/Column'
-import { AutoRow } from 'components/Row'
-import { ExternalLink } from 'theme'
-
-const parser = new UAParser(window.navigator.userAgent)
-
-const userAgent = parser.getResult()
-
-const FallbackWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  z-index: 1;
-`
-
-const BodyWrapper = styled.div<{ margin?: string }>`
-  padding: 1rem;
-  margin: auto;
-  padding: 18px 24px;
-`
-
-const CodeBlockWrapper = styled.div`
-  overflow: auto;
-  white-space: pre;
-`
-
-const LinkWrapper = styled.div`
-  margin: auto;
-`
+import FallbackView from 'components/ErrorBoundary/FallbackView'
 
 type ErrorBoundaryState = {
   error: Error | null
@@ -53,6 +18,7 @@ export default class ErrorBoundary extends React.Component<PropsWithChildren<unk
   }
 
   componentDidCatch(error: Error, errorInfo: ErrorInfo) {
+    console.log('error.name', error.name)
     if (error.name === 'ChunkLoadError' || /Loading .*?chunk .*? failed/.test(error.message)) {
       const e = new Error(`[ChunkLoadError] ${error.message}`)
       e.name = 'ChunkLoadError'
@@ -73,88 +39,9 @@ export default class ErrorBoundary extends React.Component<PropsWithChildren<unk
     const { error } = this.state
 
     if (error !== null) {
-      const encodedBody = encodeURIComponent(issueBody(error))
-
-      return (
-        <FallbackWrapper>
-          <BodyWrapper>
-            <AutoColumn gap={'md'}>
-              <Text textAlign="center">
-                <Trans>Oops! Something went wrong</Trans>
-              </Text>
-              <CodeBlockWrapper>
-                <code>
-                  <Text fontSize={10}>{error.stack}</Text>
-                </code>
-              </CodeBlockWrapper>
-              <AutoRow>
-                <LinkWrapper>
-                  <ExternalLink
-                    id="create-github-issue-link"
-                    href={`https://github.com/KyberNetwork/kyberswap-interface/issues/new?assignees=&labels=bug&body=${encodedBody}&title=${encodeURIComponent(
-                      `Crash report: \`${error.name}${error.message && `: ${error.message}`}\``,
-                    )}`}
-                    target="_blank"
-                  >
-                    <Text fontSize={16}>
-                      <Trans>Create an issue on GitHub</Trans>
-                      <span>↗</span>
-                    </Text>
-                  </ExternalLink>
-                </LinkWrapper>
-              </AutoRow>
-
-              <ButtonPrimary
-                margin="auto"
-                width="fit-content"
-                onClick={() => {
-                  localStorage.clear()
-                  window.location.reload()
-                }}
-              >
-                Refresh
-              </ButtonPrimary>
-            </AutoColumn>
-          </BodyWrapper>
-        </FallbackWrapper>
-      )
+      return <FallbackView error={error} />
     }
+
     return this.props.children
   }
-}
-
-function issueBody(error: Error): string {
-  const deviceData = userAgent
-  return `## URL
-
-${window.location.href}
-
-${
-  error.name &&
-  `## Error
-
-\`\`\`
-${error.name}${error.message && `: ${error.message}`}
-\`\`\`
-`
-}
-${
-  error.stack &&
-  `## Stacktrace
-
-\`\`\`
-${error.stack}
-\`\`\`
-`
-}
-${
-  deviceData &&
-  `## Device data
-
-\`\`\`json
-${JSON.stringify(deviceData, null, 2)}
-\`\`\`
-`
-}
-`
 }
